pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "eurosat_classifier"
version = "0.0.1"
description = "eurosat_classifier"
authors = [
  { name = "team10", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi==0.115.6",
    "torch==2.5.0",
    "torchvision==0.20.0",
    "timm==1.0.14",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "pylabels>=1.2.1",
    "kagglehub>=0.4.0",
    "hydra-core",
    "pytorch-lightning"
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "loguru",
    "matplotlib"
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
